Change loading of data_entry_helper
If called too early in the bootsrap process, the iform module is not
yet loaded and so the data_entry_helper class is not available.
This issue was not apparent until a live installation on a
Pantheon-hosted site. It is unclear why the problem manifested
itself but it might relate to cache settings.